requirements.txt
youtube-dl
yfinance
beautifulsoup4
python-dotenv
praw
luhn
dnspython
google-api-client
google-api-python-client
requests
aiohttp
toml
